@@ -5,40 +5,40 @@
   "author": "Ross Zurowski <ross@rosszurowski.com>",
   "dependencies": {
     "clsx": "^1.2.1",
-    "next": "^14.2.7",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-wrap-balancer": "^1.0.0",
-    "rehype-pretty-code": "^0.6.0",
-    "shiki": "^0.12.1"
+    "next": "^15.0.3",
+    "react": "^19.0.0-rc.1",
+    "react-dom": "^19.0.0-rc.1",
+    "shiki": "^1.24.0"
   },
   "devDependencies": {
-    "@types/node": "^18.11.18",
-    "@types/react": "^18.0.26",
-    "@types/react-dom": "^18.0.10",
+    "@content-collections/core": "^0.7.3",
+    "@content-collections/mdx": "^0.2.0",
+    "@content-collections/next": "^0.2.4",
+    "@types/node": "^22.10.1",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
     "@typescript-eslint/eslint-plugin": "^5.48.0",
     "@typescript-eslint/parser": "^5.48.0",
     "autoprefixer": "^10.4.13",
-    "contentlayer": "^0.3.3",
     "eslint": "8.31.0",
     "eslint-config-next": "^13.2.0",
     "eslint-config-prettier": "^8.6.0",
     "excerpt-html": "^1.2.2",
     "feed": "^4.2.2",
-    "globby": "^13.1.3",
-    "next-contentlayer": "^0.3.3",
-    "next-plausible": "^3.7.0",
+    "globby": "^14.0.2",
+    "next-plausible": "^3.12.4",
     "postcss": "^8.4.21",
     "prettier": "^2.8.3",
     "prettier-plugin-organize-imports": "^3.2.2",
     "prettier-plugin-tailwindcss": "^0.2.1",
-    "rehype-autolink-headings": "^6.1.1",
-    "rehype-external-links": "^2.0.1",
-    "rehype-slug": "^5.1.0",
-    "remark-gfm": "^3.0.1",
-    "remark-smartypants": "^2.0.0",
-    "tailwindcss": "^3.2.4",
-    "typescript": "^4.9.4"
+    "rehype-autolink-headings": "^7.1.0",
+    "rehype-external-links": "^3.0.0",
+    "rehype-pretty-code": "^0.14.0",
+    "rehype-slug": "^6.0.0",
+    "remark-gfm": "^4.0.0",
+    "remark-smartypants": "^3.0.2",
+    "tailwindcss": "^3.4.15",
+    "typescript": "^5.7.2"
   },
   "postcss": {
     "plugins": {
