@@ -1,7 +1,6 @@
-import cx from "clsx"
-import { getMDXComponent } from "next-contentlayer/hooks"
-import Img from "next/image"
-import { ComponentPropsWithoutRef, ReactNode } from "react"
+import { MDXContent } from "@content-collections/mdx/react"
+import NextImage from "next/image"
+import { ReactNode } from "react"
 
 const components = {
   Image,
@@ -12,14 +11,18 @@ const components = {
  * Markdown renders MDX content into components.
  */
 export default function Markdown({ code }: { code: string }) {
-  const MDXContent = getMDXComponent(code)
-  return <MDXContent components={components} />
+  return <MDXContent code={code} components={components} />
 }
 
 type ImageProps = {
+  className?: string
   children?: string
   caption?: string
-} & ComponentPropsWithoutRef<typeof Img>
+  src: string
+  alt: string
+  width: number
+  height: number
+}
 
 /**
  * Image contains an image for a post. It applies standard styling, responsive
@@ -32,19 +35,36 @@ type ImageProps = {
  *     <Image>This is a caption **with markdown**</Image>
  *
  */
-function Image(props: ImageProps) {
-  const { children, className, caption: rawCaption, ...rest } = props
+function Image({
+  children,
+  className,
+  caption: rawCaption,
+  width,
+  height,
+  src,
+  alt,
+  ...rest
+}: ImageProps) {
   const caption = rawCaption || children
 
   return (
-    <figure>
-      {/* eslint-disable-next-line jsx-a11y/alt-text */}
-      <Img
-        {...rest}
-        className={cx("bg-stone-200", className)}
-        sizes="(min-width: 48em) 50vw, 100vw"
-        quality={80}
-      />
+    <figure className={className}>
+      <div
+        className="bg-stone-200"
+        style={{
+          aspectRatio: `${width} / ${height}`,
+        }}
+      >
+        <NextImage
+          {...rest}
+          src={src}
+          alt={alt}
+          width={width}
+          height={height}
+          sizes="(min-width: 48em) 50vw, 100vw"
+          quality={80}
+        />
+      </div>
       {caption && <figcaption>{caption}</figcaption>}
     </figure>
   )
